requirements.txt
certifi==2021.5.30
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
fitbit==0.3.1
idna==2.10
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.0
more-itertools==8.8.0
oauthlib==3.1.1
portend==2.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
tempora==4.1.1
urllib3==1.26.6
zc.lockfile==2.0
